Revert "Revert "debug build""
This reverts commit d8e4fae1304dffa52ac859f7993e28230d863d55.

+++ fetch_cord/out.py
@@ -172,8 +172,11 @@ def neofetch(loop):
 
     elif not neofetchwin:
         filepath = "/tmp/out.txt"
-        with open(filepath, 'w') as f:
-            print(baseinfo, file=f)
+        if args.config:
+            filepath = args.config
+        if not args.config:
+            with open(filepath, 'w') as f:
+                print(baseinfo, file=f)
         with open(filepath, "rt") as f:
             for line in f:
                 if line.find(cpu) != -1:
@@ -228,7 +231,7 @@ def neofetch(loop):
                     cirrusgpuline.append(line.rstrip('\n'))
 
     try:
-        if os.path.isfile(filepath):
+        if os.path.isfile(filepath) and not args.config:
             os.remove(filepath)
     except FileNotFoundError:
         pass
